# src/integration-test/resources/OpenApiSpecificationIT/openapi.yaml
openapi: 3.0.1
info:
  title: api-contactsettings
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "2.0"
servers:
  - url: http://localhost:57894
    description: Generated server url
...
paths:
...
  /{municipalityId}/delegates/{id}/filters:
    post:
      tags:
        - Delegates
      summary: Create delegate filter
      operationId: create_2
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Delegate ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Filter"
        required: true
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/delegates/{id}/filters/{filterId}:
    get:
      tags:
        - Delegates
      summary: Read delegate filter
      operationId: read_1
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Delegate ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        - name: filterId
          in: path
          description: Delegate filter ID
          required: true
          schema:
            type: string
          example: b95eb1ed-0561-49f2-a7dc-5b8bc0411778
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filter"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Filter"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - Delegates
      summary: Delete delegate filter
      operationId: delete_1
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Delegate ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        - name: filterId
          in: path
          description: Delegate filter ID
          required: true
          schema:
            type: string
          example: b95eb1ed-0561-49f2-a7dc-5b8bc0411778
      responses:
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
    patch:
      tags:
        - Delegates
      summary: Update delegate filter
      operationId: update_1
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Delegate ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        - name: filterId
          in: path
          description: Delegate filter ID
          required: true
          schema:
            type: string
          example: b95eb1ed-0561-49f2-a7dc-5b8bc0411778
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Filter"
        required: true
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filter"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Filter"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/delegates/{id}:
    get:
      tags:
        - Delegates
      summary: Read delegate
      operationId: read_2
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Delegate ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Delegate"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Delegate"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - Delegates
      summary: Delete delegate
      operationId: delete_2
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Delegate ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
...
components:
  schemas:
...
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Filter:
      required:
        - rules
      type: object
      properties:
        id:
          type: string
          description: ID of the filter
          readOnly: true
          example: 5d8403b1-1bf0-4cb1-b39e-c7c504d501a1
        alias:
          type: string
          description: The filter alias
          example: My filter for delegating messages to a friend
        channel:
          type: string
          description: The channel that created this filter.
          example: Sundsvalls Energi
        created:
          type: string
          description: Timestamp when filter was created
          format: date-time
          readOnly: true
          example: 2020-08-31T01:30:00+02:00
        modified:
          type: string
          description: Timestamp when filter was last modified
          format: date-time
          readOnly: true
          example: 2020-08-31T01:30:00+02:00
        rules:
          type: array
          description: |-
            The filter rules.
            If more than one rule exists, there will be an implicit AND-condition between the rules.
            I.e. all rules must evaluate to true in order to pass the filter.
          items:
            $ref: "#/components/schemas/Rule"
      description: Filter model
    Operator:
      type: string
      description: Operator model
      example: EQUALS
      enum:
        - EQUALS
        - NOT_EQUALS
    Rule:
      required:
        - attributeName
        - attributeValue
        - operator
      type: object
      properties:
        attributeName:
          type: string
          description: The attribute name to apply the filter rule on
          example: facilityId
        operator:
          $ref: "#/components/schemas/Operator"
        attributeValue:
          type: string
          description: The attribute value to apply the filter rule on
          example: "12345678"
      description: |
        Rule model.

        Here are some rule examples.

        A rule that mathches everything. The MATCH_ALL_RULE:
        {
        "attributeName": "*",
        "operator": "EQUALS",
        "attributeValue": "*"
        }

        A rule that matches a single attribute:
        {
        "attributeName": "someAttributeName",
        "operator": "EQUALS",
        "attributeValue": "theValue"
        }

        A rule that matches everything but a single attribute (i.e. a negation of the rule above):
        {
        "attributeName": "someAttributeName",
        "operator": "NOT_EQUALS",
        "attributeValue": "theValue"
        }
...
    Delegate:
      type: object
      properties:
        id:
          type: string
          description: ID of the delegate
          readOnly: true
          example: 0d64c132-3aea-11ec-8d3d-0242ac130003
        principalId:
          type: string
          description: Contact setting ID of the delegate principal (owner)
          readOnly: true
          example: 0d64c132-3aea-11ec-8d3d-0242ac130003
        agentId:
          type: string
          description: Contact setting ID of the delegate agent (performer)
          readOnly: true
          example: 0d64c132-3aea-11ec-8d3d-0242ac130003
        created:
          type: string
          description: Timestamp when delegate was created
          format: date-time
          readOnly: true
          example: 2020-08-31T01:30:00+02:00
        modified:
          type: string
          description: Timestamp when delegate was last modified
          format: date-time
          readOnly: true
          example: 2020-08-31T01:30:00+02:00
        filters:
          type: array
          description: |-
            The delegate filters.
            If more than one filter exists, there will be an implicit OR-condition between the filters.
            I.e. at least one filter must evaluate to true in order to delegate anything.
            At least one filter must exist on a delegate. If everything should be delegated, a MATCH_ALL_RULE can be used (see: Rule-model)
          items:
            $ref: "#/components/schemas/Filter"
      description: Delegate model
      readOnly: true
  securitySchemes: { }